requirements.txt
# Core web framework
fastapi==0.115.14
uvicorn==0.35.0
python-multipart

# Streamlit for frontend
streamlit==1.46.1

# PDF processing
PyMuPDF==1.26.3
pytesseract==0.3.13
Pillow==11.3.0

# Image processing and OCR
opencv-python==4.12.0.88
numpy

# LangChain and AI/ML
langchain==0.3.26
langchain-google-genai==2.1.6
langchain-community==0.3.26
langchain-core==0.3.67

# Vector store
faiss-cpu==1.11.0

# Environment and configuration
python-dotenv==1.1.1

# HTTP requests
requests

# Data processing
pandas==2.3.0

# Type hints and validation
pydantic==2.11.7

# Additional core dependencies (without strict version constraints)
accelerate
aiofiles
aiohttp
aiosignal
altair
annotated-types
anyio
attrs
backoff
bcrypt
beautifulsoup4
blinker
build
cachetools
certifi
cffi
chardet
charset-normalizer
Chroma
chromadb
click
colorama
coloredlogs
cryptography
dataclasses-json
distro
durationpy
emoji
filelock
filetype
flatbuffers
frozenlist
fsspec
gitdb
GitPython
google-ai-generativelanguage
google-api-core
google-auth
googleapis-common-protos
greenlet
grpcio
grpcio-status
h11
html5lib
httpcore
httptools
httpx
httpx-sse
huggingface-hub
humanfriendly
idna
importlib_metadata
importlib_resources
Jinja2
jiter
joblib
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
kubernetes
langchain-chroma
langchain-huggingface
langchain-openai
langchain-text-splitters
langdetect
langsmith
lxml
Markdown
markdown-it-py
MarkupSafe
marshmallow
mdurl
opentelemetry-api
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
orjson
overrides
packaging
posthog
propcache
proto-plus
protobuf
psutil
pyarrow
pyasn1
pyasn1_modules
pybase64
pycparser
pydantic-settings
pydantic_core
pydeck
Pygments
pypdf
PyPDF2
PyPika
pyproject_hooks
pyreadline3
python-dateutil
python-iso639
python-magic
safetensors
scikit-learn
scipy
sentence-transformers
setuptools
shellingham
six
smmap
sniffio
soupsieve
SQLAlchemy
starlette
sympy
tenacity
threadpoolctl
tiktoken
tokenizers
toml
torch
tornado
tqdm
transformers
typer
typing-inspect
typing-inspection
typing_extensions
tzdata
unstructured
unstructured-client
urllib3
watchdog
watchfiles
webencodings
websocket-client
websockets
wrapt
yarl
zipp
zstandard

# Optional: Redis (if you plan to use it)
# redis 
langgraph 
